+++ src/utils/AppUtils.js
@@ -1,19 +1,19 @@
 // @flow
 
-import ENV_URLS from '../constants/EnvUrls';
-import EnvTypes from '../constants/EnvTypes';
-import Logger from '../Logger';
-import ParticipantDataTypes from '../constants/ParticipantDataTypes';
-import { isNonEmptyString } from '../LangUtils';
-import { isValidUUID } from '../ValidationUtils';
+import ENV_URLS from './constants/EnvUrls';
+import EnvTypes from './constants/EnvTypes';
+import Logger from './Logger';
+import ParticipantDataTypes from './constants/ParticipantDataTypes';
+import { isNonEmptyString } from './LangUtils';
+import { isValidUUID } from './ValidationUtils';
 import {
   CHRONICLE,
   DATA,
   FILE_TYPE,
   PARTICIPANT,
   STUDY,
-} from '../constants/UrlConstants';
-import type { ParticipantDataType } from '../constants/ParticipantDataTypes';
+} from './constants/UrlConstants';
+import type { ParticipantDataType } from './constants/ParticipantDataTypes';
 
 const LOG = new Logger('AppApi');
 const { LOCAL, PRODUCTION, STAGING } = EnvTypes;
