+++ src/utils/DateUtils.js
@@ -10,16 +10,6 @@ const getDateTimeFromIsoDate = (isoDate :string) => {
   return null;
 };
 
-// format: Tuesday, October 14, 1983
-const getFullDateFromIsoDate = (isoDate :string) => {
-  const date :DateTime = DateTime.fromISO(isoDate);
-  if (date.isValid) {
-    return date.toLocaleString(DateTime.DATE_HUGE);
-  }
-  return null;
-};
-
 export {
   getDateTimeFromIsoDate,
-  getFullDateFromIsoDate
 };
